1>>> float_dtype is set to Double even for CUDA.  2>>> SplineNet and SplineNet_ are now generalized to accept *spline_shape*, which allows to have a field of spline networks. 3>>> the step method in Model.fit can now handle a NaN value in loss; by not following the step. 4>>> One important change was made in 0b97e0dd35390959db09a038ad6a1298201c6a95 commit, in which the beta of Softmax in nn/scalar/modules.py and nn/scalar/coupling_.py are set log(2); we now change the initial values for weights_d to zero.

+++ src/_global_settings.py
@@ -7,8 +7,10 @@ import numpy as np
 
 if torch.cuda.is_available():
     torch_device = 'cuda'
-    float_dtype = torch.float32  # np.float32, single
-    float_tensortype = torch.cuda.FloatTensor
+    # float_dtype = torch.float32  # np.float32, single
+    # float_tensortype = torch.cuda.FloatTensor
+    float_dtype = torch.float64  # np.float64, double
+    float_tensortype = torch.cuda.DoubleTensor
 else:
     torch_device = 'cpu'
     float_dtype = torch.float64  # np.float64, double
